Update org-mode template examples
Updated the elisp and src examples so that they do not use `indent-according-to-tab`. When the header of the source block exists but the footer does not yet exist org-mode greedily looks for the next `#+end_src`, even if a `#+begin_src` precedes it. If it ends that `#+end_src` it then treats all text between tempel's freshly inserted src header and that existing footer as source doe and attempts to indent and possible format it accordingly. The results will vary by the source language, but for elisp it's indenting all text and then whenever it finds a `#` it prepends a `,`.

This change introduces no degradation to the template examples, but it does work with org-mode as is.