Ensure the order of overlays is maintained as an invariant.
This is done via the function tempel--ensure-ov-invariant, which
should be called after every field and form overlay motions. The
invariant is maintained by making sure the boundaries of every overlay
that was added to the state before the current overlay are clamped to
the start of the current overlay, and likewise for overlays that were
added later. Since overlays are added to the state in order, this
ensures the invariant.